dump environment var

+++ install.sh
@@ -6,7 +6,7 @@ rm -fr /tmp/SimulationEngineInstall
 
 cp -r /tmp/SimulationEngine /tmp/SimulationEngineInstall
 
-ENVIRONMENT=dev
+ENVIRONMENT="dev"
 export ENVIRONMENT
 
 cd /tmp/SimulationEngineInstall
@@ -14,7 +14,9 @@ cd /tmp/SimulationEngineInstall
 pip3 install -r requirements.txt
 
 # kill running python process
-kill -9 `ps -eaf | grep simulation_manager_main | awk '{print $2}'`
+if pgrep simulation_manager_main; then kill -9 `ps -eaf | grep simulation_manager_main | awk '{print $2}'`; fi
+
+echo "ENVIRONMENT=${ENVIRONMENT}"
 
 # start python process
 python3 simulation_manager_main.py > /dev/null 2> /dev/null < /dev/null &
